fix(client-forms): .int() в weekDays-override для всех 3 task-форм
Зеркальный fix к 513f6cd (server schema). Client-side формы
(CreateTask / EditTask / DuplicateTaskDialog) extend'или
insertTaskSchema, но переопределяли weekDays через
`z.array(z.number()).nullable().optional()` БЕЗ .int(). Это снимало
guard который только что добавили на сервере.

Раньше последовательность была:
  1. UI form пропускает float (не должно случиться через checkbox-
     UI, но defensive)
  2. Server отвергает с 400 ZodError
  3. UI показывает раздражающий «Ошибка валидации»

Теперь client отвергает локально с понятным «Неверный формат» сразу
в форме без round-trip к серверу.

Аналогичная защита от corrupted localStorage / browser auto-fill
с float значениями.

(typecheck зелёный, 1105 тестов зелёные)

+++ client/src/components/DuplicateTaskDialog.tsx
@@ -38,7 +38,8 @@ import { useToast } from "@/hooks/use-toast";
 const formSchema = insertTaskSchema.extend({
   workerId: z.string().optional().transform(val => val ? parseInt(val, 10) : undefined),
   requiresPhoto: z.boolean().optional().default(false),
-  weekDays: z.array(z.number()).nullable().optional(),
+  // .int() симметрично с insertTaskSchema (защита от float'ов).
+  weekDays: z.array(z.number().int()).nullable().optional(),
   monthDay: z.string().optional().transform(val => val ? parseInt(val, 10) : null),
   isRecurring: z.boolean().optional().default(true),
   price: z.string().optional().transform(val => val ? parseInt(val, 10) : 0),
